requirements.txt
streamlit>=1.28.0
python-docx>=0.8.11
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
